fix PortForwardingMacApp
Summary:
Since using the new electron-builder all of our source files are bundled in a zip-like ASAR bundle for distribution. This also happened for the PortForwardingMacApp, so the app couldn't be launched anymore.

The app is now added to the `asarUnpack` config, which unpacks files that can't be part of the bundle for some reason. This again, broke code signing, because the unbundled app could not be code signed. The problem was the symlinks in the applications bundle were replaced by copies of their destination.

Moreover, the instance is killed when Sonar closes. before every time Sonar did reload, a new instance was spawned.

Reviewed By: emilsjolander

Differential Revision: D7727628

fbshipit-source-id: 7a78ee6fe79424c3a9dc7bac31feee1895578922