[gradle] Set up Maven packaging and Bintray upload
Summary:
This make use of the the setup we have at Litho, which should make
it easier to keep them in sync.

It creates proper source jars, AARs (without bundling in unwanted
transitive deps) and POMs.

Test Plan:
```
./gradlew :android:bintrayUpload -info -PdryRun=true
```

This looks good to me, but this is one of those things you never quite
know if they work until you test them in prod.