[android] Remove NDK build from fbjni sub-project
Summary:
We're already building this as dependency via CMake of sonar itself and
bundle the resulting `.so` files in there.

This project only contains the java files so we can ship them as
separate JAR.

Test Plan:
Published the result of this as `0.0.8` and built the sample app with
it.