Remove MySonarPlugin from getting-started
I think the getting-started page should be a no-nonsense quick set of steps to get you up and running so you can try out flipper.

With it referencing the non existent "MySonarPlugin", you can't just follow the steps for it to work. You'd have to either write a plugin which would take a while, or look at the sample apps on your own to see how to add a real plugin. It seems far simpler just to give a working example.