Remove MySonarPlugin from getting-started (#240)
Summary:
I think the getting-started page should be a no-nonsense quick set of steps to get you up and running so you can try out flipper.

With it referencing the non existent "MySonarPlugin", you can't just follow the steps for it to work. You'd have to either write a plugin which would take a while, or look at the sample apps on your own to see how to add a real plugin. I think it will be far more helpful just to give a working example.
Pull Request resolved: https://github.com/facebook/flipper/pull/240

Reviewed By: passy

Differential Revision: D9378196

Pulled By: jknoxville

fbshipit-source-id: 18ae4a39d09c7baa17adb7f7a71fd766e9bc8e51