iOS: Add very basic diagnostic screen
Summary:
Adds an extremely barebones in-app screen for showing sonar diagnostics.
The UI is laughable, but it's wired up. Leaving proper design to a later commit.

Like in the android case, state is currently passed as a string, but will become something much more structured after this is in place.

Reviewed By: priteshrnandgaonkar

Differential Revision: D9218766

fbshipit-source-id: 4889ed79b928056a67b1e8cb8a40a9bd52e084f1