Build Sample app rather than just building the sample pods (#251)
Summary:
Currently travis just builds the pods which sample app depends on. The test case which the current setup misses it that it will fail to check if the required header files are exposed to public. The master is broken recently because a header file is not exposed publicly and sample app fails to build, whereas all the dependencies like Sonar and SonarKit build successfully. This PR updates the travis, so that it builds sample app rather than the dependent pods.
Pull Request resolved: https://github.com/facebook/flipper/pull/251

Reviewed By: jknoxville

Differential Revision: D9465234

Pulled By: priteshrnandgaonkar

fbshipit-source-id: 42ab7210d7fa5a3dc2927efee4b4dee95f15c6cd