Remove accessibility panel from Layout Inspector in non-accessibility mode
Summary: Removes the Accessibility panel from the layout inspector since all of these properties are shown (and logged) in the Accessibility Inspector along with other properties that may be useful. This will make it easier to track accessibility use and also will help more people discover the accessibility feature when they need it.

Reviewed By: blavalla

Differential Revision: D9523532

fbshipit-source-id: f6d3a745980ec631ec8bf2e59017382f4c2edd5a