Enabling fullscreen mode on macos (#259)
Summary:
I wanted to use Flipper/Sonar in a fullscreen/split screen environment end was surprised the feature was disabled. When I checked the blame it looks like it's been like this from the get go, so there might not even be a real reason to have it disabled. Please correct me if I'm wrong.

The default value for `fullsreenable` is true. Removing setting the
parameter reenables the feature.

- Flipped the switch, saw it go fullscreen
- Also works in split screens
Pull Request resolved: https://github.com/facebook/flipper/pull/259

Reviewed By: danielbuechele

Differential Revision: D9630068

Pulled By: passy

fbshipit-source-id: 650e54895dc596090e2edcf2802bb0020f207ecc