Don't use vendorID as deviceId
Summary:
Device Id needs to be the same for multiple apps on the same device. This isn't true for this attribute, which is specific to the vendor of the app.
Apple has deprecated the API that gives you the device Id, so we can't discover it in the SDK alone.

However, this isn't a problem because we have a similar issue on android, and now Flipper desktop gives each device it's own deviceId during certificate exchange.
Removing this incorrect one so it's never used as a fallback, which would cause flipper to think the app is on a different device than the one it's running on.

Reviewed By: danielbuechele

Differential Revision: D9700552

fbshipit-source-id: 5467f6cb7d1f941af9df09fc6a03dbe34e6575e5